Create Solidity Contract
In this contract, we define the EtherToTokenSwapper contract that implements the ERC20Swapper interface. The swapEtherToToken function takes the address of the ERC-20 token and the minimum amount of tokens required as input.

Inside the function, we create an instance of the ERC-20 token contract using the provided token address. Then, we calculate the amount of tokens to transfer based on the exchange rate or any other logic. In this example, we assume a simple 1:1 exchange rate between Ether and tokens.

We then perform some checks to ensure that the calculated token amount is equal to or greater than the minimum amount required. If the check passes, we use the transfer function of the ERC20 token contract to transfer the tokens to the message sender (the person who called the swapEtherToToken function).

Finally, we return the token amount that was transferred.